+++ src/Sql/Sql/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 * Added `DatabaseFormat` and `PricingModel` parameters to `New-AzSqlInstance`, `Set-AzSqlInstance`
+* Added a new cmdlet for Azure SQL Managed Instance refresh external governance status
+  - 'Invoke-AzSqlInstanceExternalGovernanceStatusRefresh'
+* Updated `Get-AzSqlInstance` to support returning the `ExternalGovernanceStatus` property
 
 ## Version 4.13.0
 * Fixed `Set-AzSqlDatabaseFailoverGroup` when going from multi-secondary to single secondary